+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [FormElement](src/formElement.ts)
   - Increased height of `button[type=submit]`
   - Added callback `$onSubmitEnd` & event `$submitEnd`
+  - Web Accessibility: auto-tie with closest heading via aria-labelledby
 - **Text based controls**
   - Added css variable `--ctrl-label-active-pos` to simplify customization of label position
 - [SelectControl](src/controls/select.ts). [SelectManyControl](src/controls/selectMany.ts). Improved handling promise rejection in `$options.items`